pyproject.toml
[project]
name = "simplify-text"
version = "0.1.0"
description = "AI-based text simplifier using Azure OpenAI for improved readability"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["ai", "text-processing", "nlp", "simplification", "azure-openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=2.21.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Repository = "https://github.com/your-username/simplify-text"
Documentation = "https://github.com/your-username/simplify-text#readme"

[project.scripts]
simplify-text = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "main.py",
    "test_simplifier.py",
]

[dependency-groups]
dev = []
